split fixture into more digestable chunks
